requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
SQLAlchemy>=2.0.36
psycopg2-binary>=2.9.10
pydantic>=2.9.0
pydantic[email]>=2.9.0
python-dotenv>=1.0.0
alembic>=1.13.0
pytest>=8.3.0
httpx>=0.27.0
pytest-asyncio>=0.24.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.0,<5.0.0
python-multipart>=0.0.12
email-validator>=2.0.0
